+++ tests/testthat/test_asym_gee.R
@@ -140,9 +140,9 @@ if (requireNamespace("plm")) {
 
 # tidiers -----------------------------------------------------------------
 context("asym_gee tidiers")
-if (requireNamespace("broom")) {
-  expect_is(broom::tidy(wb <- asym_gee(wks ~ lag(union) + lag(lwage),
+if (requireNamespace("broom.mixed")) {
+  expect_is(generics::tidy(wb <- asym_gee(wks ~ lag(union) + lag(lwage),
                                        data = wages), conf.int = TRUE), 
             "data.frame")
-  expect_is(broom::glance(wb), "data.frame")
+  expect_is(generics::glance(wb), "data.frame")
 }