+++ tests/testthat/test_wbgee.R
@@ -203,11 +203,11 @@ if (requireNamespace("plm")) {
 
 # tidiers -----------------------------------------------------------------
 context("wbgee tidiers")
-if (requireNamespace("broom")) {
-  expect_is(broom::tidy(wb <- wbgee(wks ~ lag(union) + lag(lwage),
+if (requireNamespace("broom.mixed")) {
+  expect_is(generics::tidy(wb <- wbgee(wks ~ lag(union) + lag(lwage),
                                        data = wages), conf.int = TRUE), 
             "data.frame")
-  expect_is(broom::glance(wb), "data.frame")
+  expect_is(generics::glance(wb), "data.frame")
 }
 
 # predictions ----------------------------------------------------------------